+++ packages/coast_audio/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `AudioFilterNode` which can be used to apply filters easily.
 - Fixed `FrameRingBuffer.copyTo()` returns number of bytes copied instead of number of frames copied.
 - `RingBuffer` and `FrameRingBuffer` will be disposed automatically when they are no longer used.
+- `MixerNode` now supports multiple SampleFormat.int32 and SampleFormat.uint8.
 
 ### Breaking Changes